move confirm delete modal to shared
